fix: The service and repository interfaces are separated into a file
