+++ internal/service/warehouse.go
@@ -5,27 +5,10 @@ import (
 	"sort"
 
 	"github.com/varobledo_meli/W17-G10-Bootcamp.git/internal/mappers"
-	"github.com/varobledo_meli/W17-G10-Bootcamp.git/internal/repository"
 	"github.com/varobledo_meli/W17-G10-Bootcamp.git/internal/validators"
 	"github.com/varobledo_meli/W17-G10-Bootcamp.git/pkg/api"
 	"github.com/varobledo_meli/W17-G10-Bootcamp.git/pkg/models/warehouse"
 )
-
-type WarehouseService interface {
-	Create(ctx context.Context, w warehouse.Warehouse) (*warehouse.Warehouse, *api.ServiceError)
-	FindAll(ctx context.Context) ([]warehouse.Warehouse, *api.ServiceError)
-	FindById(ctx context.Context, id int) (*warehouse.Warehouse, *api.ServiceError)
-	Update(ctx context.Context, id int, patch warehouse.WarehousePatchDTO) (*warehouse.Warehouse, *api.ServiceError)
-	Delete(ctx context.Context, id int) *api.ServiceError
-}
-
-type WarehouseDefault struct {
-	rp repository.WarehouseRepository
-}
-
-func NewWarehouseDefault(rp repository.WarehouseRepository) *WarehouseDefault {
-	return &WarehouseDefault{rp: rp}
-}
 
 func (s *WarehouseDefault) Create(ctx context.Context, w warehouse.Warehouse) (*warehouse.Warehouse, *api.ServiceError) {
 	ok, err := s.rp.Exist(ctx, w.WarehouseCode)
